add immutable BaseToken
