+++ contracts/QuoteTokenPerpdex.sol
@@ -3,5 +3,9 @@ pragma solidity 0.7.6;
 import { VirtualTokenPerpdex } from "./VirtualTokenPerpdex.sol";
 
 contract QuoteTokenPerpdex is VirtualTokenPerpdex {
-    constructor(string memory name, string memory symbol) VirtualTokenPerpdex(name, symbol) {}
+    constructor(
+        string memory name,
+        string memory symbol,
+        address recipient
+    ) VirtualTokenPerpdex(name, symbol, recipient) {}
 }
