+++ contracts/VirtualTokenPerpdex.sol
@@ -5,9 +5,11 @@ import { ERC20 } from "@openzeppelin/contracts/token/ERC20/ERC20.sol";
 import { IVirtualToken } from "./interface/IVirtualToken.sol";
 
 contract VirtualTokenPerpdex is ERC20 {
-    constructor(string memory name, string memory symbol) public ERC20(name, symbol) {}
-
-    function mintMaximumTo(address recipient) external {
+    constructor(
+        string memory name,
+        string memory symbol,
+        address recipient
+    ) public ERC20(name, symbol) {
         _mint(recipient, type(uint256).max);
     }
 }
