+++ contracts/interface/IBaseTokenNew.sol
@@ -2,30 +2,6 @@
 pragma solidity >=0.7.6;
 
 interface IBaseTokenNew {
-    // Do NOT change the order of enum values because it will break backwards compatibility
-    enum Status { Open, Paused, Closed }
-
-    event PriceFeedChanged(address indexed priceFeed);
-    event StatusUpdated(Status indexed status);
-
-    function close() external;
-
-    /// @notice Get the price feed address
-    /// @return priceFeed the current price feed
-    function getPriceFeed() external view returns (address priceFeed);
-
-    function getPausedTimestamp() external view returns (uint256);
-
-    function getPausedIndexPrice() external view returns (uint256);
-
-    function getClosedPrice() external view returns (uint256);
-
-    function isOpen() external view returns (bool);
-
-    function isPaused() external view returns (bool);
-
-    function isClosed() external view returns (bool);
-
     function shareToBalance(uint256 share) external view returns (uint256);
 
     function balanceToShare(uint256 balance) external view returns (uint256);
