+++ deploy/021_deploy_clearing_house.ts
@@ -4,19 +4,20 @@ import { parseEther } from "ethers/lib/utils"
 
 const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
     const { deployments, getNamedAccounts } = hre
-    const { deploy } = deployments
+    const { deploy, execute } = deployments
     const { deployer } = await getNamedAccounts()
 
-    const quoteToken = await deployments.get("QuoteTokenNative")
-    const uniV2Factory = await deployments.get("UniswapV2Factory")
+    const quoteTokenName = "QuoteEth"
+    const quoteTokenSymbol = "QUOTEETH"
 
+    const uniV2Factory = await deployments.get("UniswapV2Factory")
     const accountLibrary = await deployments.get("AccountLibrary")
     const takerLibrary = await deployments.get("TakerLibrary")
     const makerLibrary = await deployments.get("MakerLibrary")
 
     await deploy("ClearingHousePerpdexNew", {
         from: deployer,
-        args: [quoteToken.address, uniV2Factory.address],
+        args: [quoteTokenName, quoteTokenSymbol, uniV2Factory.address],
         libraries: {
             AccountLibrary: accountLibrary.address,
             TakerLibrary: takerLibrary.address,
